requirements.txt
# Dev environment
pip
autopep8

# Libraries
python-docx
Pillow
